+++ lib/src/activatory.dart
@@ -1,6 +1,7 @@
 import 'dart:core';
-import 'package:activatory/src/activation_context_factory.dart';
+
 import 'package:activatory/src/activation_context.dart';
+import 'package:activatory/src/activation_context_factory.dart';
 import 'package:activatory/src/backends/explicit_backend.dart';
 import 'package:activatory/src/backends/singleton_backend.dart';
 
@@ -10,15 +11,13 @@ class Activatory {
   ActivationContext _context;
   ActivationContextFactory _factory;
 
-  Activatory.custom(this._factory,[this._context]){
-    if(_context==null){
+  Activatory() : this.custom(new ActivationContextFactory());
+
+  Activatory.custom(this._factory, [this._context]) {
+    if (_context == null) {
       _context = _factory.createDefault();
     }
   }
-
-  Activatory():this.custom(new ActivationContextFactory());
-
-  T getTyped<T>({Object key}) => get(T, key: key);
 
   Object get(Type type, {Object key}) {
     var backend = _context.get(type, key: key);
@@ -26,7 +25,9 @@ class Activatory {
     return value;
   }
 
-  void override<T>(Generator<T> generator, {Object key}){
+  T getTyped<T>({Object key}) => get(T, key: key);
+
+  void override<T>(Generator<T> generator, {Object key}) {
     var backend = new ExplicitBackend(generator);
     _context.register(backend, T, key: key);
   }
@@ -40,5 +41,7 @@ class Activatory {
     _context.register(backend, type, key: key);
   }
 
-  void pinValue<T>(T value, {Object key})=> override((ctx) => value, key: key);
+  void pinValue<T>(T value, {Object key}) => override((ctx) => value, key: key);
+
+  void registerArray<T>() => _context.registerArray<T>();
 }