+++ CHANGELOG.md
@@ -18,3 +18,7 @@
 ## v0.0.8
 - Back-ends now can be registered and lookuped using any object as a key
 - useSingleton renamed to pin, useValue renamed to pinValue
+
+## v0.0.9
+- Added effortless support for primitive types arrays generating (except of enums arrays)
+- Added support for complex types array generating with explicit array registration